@@ -22,32 +22,32 @@
   "devDependencies": {
     "@types/convict": "6.1.1",
     "@types/convict-format-with-validator": "6.0.2",
-    "@types/node": "16.11.26",
-    "@typescript-eslint/eslint-plugin": "5.18.0",
-    "@typescript-eslint/parser": "5.18.0",
+    "@types/node": "18.11.13",
+    "@typescript-eslint/eslint-plugin": "5.46.0",
+    "@typescript-eslint/parser": "5.46.0",
     "chalk": "5.0.1",
-    "eslint": "8.12.0",
-    "eslint-config-htmlacademy": "5.1.2",
-    "json-server": "0.17.0",
+    "eslint": "8.29.0",
+    "eslint-config-htmlacademy": "8.0.0",
+    "json-server": "0.17.1",
     "pino-pretty": "9.1.1",
     "rimraf": "3.0.2",
-    "ts-node": "10.7.0",
-    "typescript": "4.6.3"
+    "ts-node": "10.9.1",
+    "typescript": "4.9.4"
   },
   "type": "module",
   "engines": {
-    "node": "16.13.0",
-    "npm": "8.1.0"
+    "node": "16.13.1",
+    "npm": "8.1.2"
   },
   "dependencies": {
-    "@typegoose/typegoose": "9.12.1",
+    "@typegoose/typegoose": "9.13.2",
     "convict": "6.2.3",
     "convict-format-with-validator": "6.2.0",
-    "dayjs": " 1.11.6",
+    "dayjs": "1.11.7",
     "dotenv": "16.0.3",
-    "got": "12.5.2",
+    "got": "12.5.3",
     "inversify": "6.0.1",
-    "mongoose": "6.6.5",
+    "mongoose": "6.7.2",
     "pino": "8.7.0",
     "reflect-metadata": "0.1.13"
   }
