@@ -1,8 +1,11 @@
 import { LoggerInterface } from '../common/logger/logger.interface.js';
 import { ConfigInterface } from '../common/config/config.interface.js';
-import { inject, injectable } from 'inversify';
+import {
+  inject,
+  injectable
+} from 'inversify';
 import { Component } from '../types/component.types.js';
-import { getURI } from '../utils/db.js';
+import { getDatabaseURI } from '../utils/db.js';
 import { DatabaseInterface } from '../common/database-client/database.interface.js';
 
 @injectable()
@@ -17,7 +20,7 @@ export default class Application {
     this.logger.info('App initialization...');
     this.logger.info(`Get value from env $PORT: ${this.config.get('PORT')}`);
 
-    const uri = getURI(
+    const databaseURI = getDatabaseURI(
       this.config.get('DB_USER'),
       this.config.get('DB_PASSWORD'),
       this.config.get('DB_HOST'),
@@ -25,6 +28,6 @@ export default class Application {
       this.config.get('DB_NAME'),
     );
 
-    await this.databaseClient.connect(uri);
+    await this.databaseClient.connect(databaseURI);
   }
 }
