@@ -3,7 +3,7 @@ import { CliCommandInterface } from '../cli-command/cli-command.interface.js';
 type ParsedCommand = Record<string, string[]>
 
 export default class CLIApplication {
-  private commands: {[propertyName: string]: CliCommandInterface} = {};
+  private commands: Record<string, CliCommandInterface> = {};
   private defaultCommand = '--help';
 
   private parseCommand(cliArguments: string[]): ParsedCommand {
