@@ -1,8 +1,14 @@
 import { config } from 'dotenv';
-import { inject, injectable } from 'inversify';
+import {
+  inject,
+  injectable
+} from 'inversify';
 import { ConfigInterface } from './config.interface.js';
 import { LoggerInterface } from '../logger/logger.interface.js';
-import { configSchema, ConfigSchema } from './config.schema.js';
+import {
+  configSchema,
+  ConfigSchema
+} from './config.schema.js';
 import { Component } from '../../types/component.types.js';
 
 @injectable()
