@@ -8,7 +8,7 @@ export class LoggerService implements LoggerInterface {
 
   constructor() {
     this.logger = pino();
-    this.logger.info('Logger created…');
+    this.logger.info('Logger created');
   }
 
   public debug(message: string, ...args: unknown[]): void {
